+++ ez-med-sched/Views/CalendarScreen/CalendarScreen.js
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import { Text, View } from 'react-native';
 
-const HomeScreen = () => (
+const CalendarScreen = () => (
     <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-        <Text>Home!</Text>
+        <Text>Calendar!</Text>
     </View>);
 
 
 
-export default HomeScreen;
+export default CalendarScreen;
