+++ ez-med-sched/Views/ProfileScreen/ProfileScreen.js
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import { Text, View } from 'react-native';
 
-const HomeScreen = () => (
+const ProfileScreen = () => (
     <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-        <Text>Home!</Text>
+        <Text>Profile!</Text>
     </View>);
 
 
 
-export default HomeScreen;
+export default ProfileScreen;
